requirements.txt
asgiref==3.2.3
Babel==2.8.0
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.3
django-exclusivebooleanfield==0.3.0
django-grappelli==2.14.1
django-grappelli-filters==0.2.1
django-import-export==2.0.1
django-js-asset==1.2.2
django-localflavor==2.2
django-phonenumber-field==4.0.0
djangoql==0.13.1
djangorestframework==3.11.0
et-xmlfile==1.0.1
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.3
phonenumberslite==8.11.4
ply==3.11
python-stdnum==1.13
pytz==2019.3
PyYAML==5.3
six==1.14.0
sqlparse==0.3.0
tablib==0.14.0
xlrd==1.2.0
xlwt==1.3.0
